Remove PSR7 Stream wrapper from GoogleDirveAdapter
The PSR7 stream wrapper was closing the underlying resource when it was
getting freed at the end of the function. There was no way to prevent it
from freeing the underlying resource because the free happens in the
destructor. The only sane option was to remove the stream wrapper, which
is unfortunate.

The resource needs to be kept open so the file can be mirrored to
Backblaze or any other additional storage providers.