Mirror uploads to Backblaze
I am uncomfortable with relying on one "comsumer grade" hosting solution
for files. Backblaze is a more traditional cloud storage provider, and
is extremely cheap. If all gos well, I may remove the google drive
to cut costs.

I had to use a fork of the backblaze-b2 client implementation that
retries when uploads fail, since this is apparently a regular occurance
with backblaze. Its also worth noting that the upload is not chunked,
which is sub optimal.