Fix broken redirect
When someone visits a .blend with a legacy numeric id such as /b/1 we
redirect them to the new string based ID.

This is done on before route enter, to minimize the level of
"jankyiness" observed. However, when this occurs it does not refetch
the blend file because beforeRouteEnter is not called on a parameter
change. (They *really* should be a unified method).

This means the .blend file is never set, which is a bug.